reorganized code so that tasklistitems has a test for its methods in its file.
